move createBlockCommittedEvent
