+++ blockchain/service.go
@@ -1,6 +1,9 @@
 package blockchain
 
-import "github.com/it-chain/it-chain-Engine/core/eventstore"
+import (
+	"github.com/it-chain/it-chain-Engine/core/eventstore"
+	"github.com/it-chain/midgard"
+)
 
 type BlockQueryService interface {
 	BlockQueryInnerService
@@ -32,4 +35,16 @@ func CommitBlock(block Block) error {
 	eventstore.Save(blockId, event)
 
 	return nil
+}
+
+func createBlockCommittedEvent(block Block) (*BlockCommittedEvent, error) {
+
+	aggregateId := string(block.GetSeal())
+
+	return &BlockCommittedEvent{
+		EventModel: midgard.EventModel{
+			ID: aggregateId,
+		},
+		State: Committed,
+	}, nil
 }
